Extract the capacity/curve getters to capacity_curves.py (#509, V2-09)
Move get_dcap / get_ccap / get_cap / _get_cap / get_ocv (~650 lines) verbatim
from cellreader.py to the new cellpy/readers/capacity_curves.py (the
long-standing 'deserve their own module' TODO). Functions take the CellpyCell
instance as first argument; cross-calls go through the instance so subclass
dispatch is preserved bit-for-bit. CellpyCell keeps thin delegate methods
with identical explicit signatures - public API unchanged; the curve golden
suite passes against the existing goldens (no regeneration).

cellreader.py: 5987 -> 4778 lines across this branch (-20.2%).
